@@ -1,25 +1,25 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
+import {BrowserModule} from '@angular/platform-browser';
+import {NgModule} from '@angular/core';
 
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { NavigationComponent } from './navigation/navigation.component';
-import { GrafanaComponent } from './grafana/grafana.component';
-import { BuildOverviewComponent } from './build-overview/build-overview.component';
-import {CalendarHeatmap} from "angular2-calendar-heatmap";
-import {HttpClientModule} from "@angular/common/http";
+import {AppRoutingModule} from './app-routing.module';
+import {AppComponent} from './app.component';
+import {NavigationComponent} from './navigation/navigation.component';
+import {GrafanaComponent} from './grafana/grafana.component';
+import {BuildOverviewComponent} from './build-overview/build-overview.component';
+import {HttpClientModule} from '@angular/common/http';
+import {NgApexchartsModule} from 'ng-apexcharts';
 
 @NgModule({
   declarations: [
     AppComponent,
     NavigationComponent,
     GrafanaComponent,
-    BuildOverviewComponent,
-    CalendarHeatmap,
+    BuildOverviewComponent
   ],
   imports: [
     BrowserModule,
     HttpClientModule,
+    NgApexchartsModule,
     AppRoutingModule
   ],
   providers: [],
