@@ -1,6 +1,6 @@
-import { Injectable } from '@angular/core';
-import {HttpClient} from "@angular/common/http";
-import {Observable} from "rxjs";
+import {Injectable} from '@angular/core';
+import {HttpClient} from '@angular/common/http';
+import {Observable} from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
@@ -9,8 +9,8 @@ export class BuildOverviewService {
 
   constructor(private http: HttpClient) { }
 
-  getBuildOverview(): Observable<any[]> {
-    return this.http.get<any[]>('/api/v1/jenkins/build/distribution')
+  getBuildOverview(): Observable<any> {
+    return this.http.get<any>('/api/v1/jenkins/build/distribution');
   }
 
 }
