@@ -1,13 +1,13 @@
 package de.jenkinsutil.controller;
 
-import de.jenkinsutil.payload.out.BuildDay;
+import de.jenkinsutil.payload.out.SeriesHolder;
 import de.jenkinsutil.service.BuildDistributionService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import java.io.IOException;
 import java.util.Optional;
 
 @RestController
@@ -21,8 +21,9 @@ public class BuildDistributionController {
     }
 
     @GetMapping("/distribution")
-    public ResponseEntity<List<BuildDay>> getBuildDistribution() {
-        Optional<List<BuildDay>> optional = buildDistributionService.getBuildDays();
+    public ResponseEntity<SeriesHolder> getBuildDistribution() throws IOException {
+        Optional<SeriesHolder> optional = buildDistributionService.getBuildDays();
         return optional.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.noContent().build());
     }
+
 }
