@@ -3,60 +3,65 @@ package de.jenkinsutil.payload.mapping;
 import de.jenkinsutil.payload.in.JenkinsBuild;
 import de.jenkinsutil.payload.in.JenkinsBuildDistribution;
 import de.jenkinsutil.payload.in.JenkinsJob;
-import de.jenkinsutil.payload.out.BuildDay;
-import de.jenkinsutil.payload.out.BuildDetails;
+import de.jenkinsutil.payload.out.BuildJob;
+import de.jenkinsutil.payload.out.Series;
+import de.jenkinsutil.payload.out.SeriesHolder;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
 
 import java.time.Instant;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.*;
 
 @Component
 public class BuildDistributionMapper {
 
-    public static final int MILLISECONDS_IN_SECOND = 1000;
+    public SeriesHolder convert(JenkinsBuildDistribution input) {
 
-    public List<BuildDay> convert(JenkinsBuildDistribution input) {
+        SeriesHolder holder = new SeriesHolder();
 
-        List<BuildDay> buildDays = new ArrayList<>();
 
+        Map<String, Series> seriesMap = new HashMap<>();
         List<JenkinsJob> jobs = input.getJobs();
         for (JenkinsJob job : jobs) {
 
+            String seriesName = job.getName().toLowerCase().contains("test") ? "test" : "other";
+
+            Series s = getSeriesFromMap(seriesName, seriesMap);
             List<JenkinsBuild> builds = job.getBuilds();
             for (JenkinsBuild build: builds) {
                 long timestamp = build.getTimestamp();
                 LocalDateTime buildStart = LocalDateTime.ofInstant(Instant.ofEpochMilli(timestamp), TimeZone.getDefault().toZoneId());
-                LocalDate localDate = buildStart.toLocalDate();
-                if (buildStart.isBefore(LocalDateTime.now().minusYears(1))) {
+                if (buildStart.isBefore(LocalDateTime.now().minusDays(2))) {
                     continue;
                 }
-                Optional<BuildDay> optional = buildDays.stream().filter(day -> day.getDate().isEqual(localDate)).findAny();
-                if (optional.isPresent()) {
-                    BuildDay buildDay = optional.get();
-                    BuildDetails details = new BuildDetails();
-//                    details.setJobName(job.getName() + build.getNumber());
-                    details.setSlave(StringUtils.isEmpty(build.getBuiltOn()) ? "unknown" : build.getBuiltOn());
-                    details.setBuildStartTime(buildStart);
-                    details.setDurationInSeconds(Math.toIntExact(build.getDuration() / MILLISECONDS_IN_SECOND));
-                    buildDay.addBuildDetail(details);
-                } else {
-                    BuildDay buildDay = new BuildDay();
-                    buildDay.setDate(localDate);
-                    BuildDetails details = new BuildDetails();
-//                    details.setJobName(job.getName() + build.getNumber());
-                    details.setSlave(StringUtils.isEmpty(build.getBuiltOn()) ? "unknown" : build.getBuiltOn());
 
-                    details.setBuildStartTime(buildStart);
-                    details.setDurationInSeconds(Math.toIntExact(build.getDuration() / MILLISECONDS_IN_SECOND));
-                    buildDay.addBuildDetail(details);
-                    buildDays.add(buildDay);
-                }
+                BuildJob buildJob = new BuildJob();
+                buildJob.setNodeName(getBuildOn(build));
+                buildJob.setStart(build.getTimestamp());
+                buildJob.setEnd(build.getTimestamp() + build.getDuration());
+                s.addBuildJob(buildJob);
             }
+
+
         }
 
-        return buildDays;
+        List<Series> series = new ArrayList<>(seriesMap.values());
+        holder.setSeries(series);
+        return holder;
     }
+
+    private Series getSeriesFromMap(String key, Map<String, Series> seriesMap) {
+        if (!seriesMap.containsKey(key)) {
+            Series series = new Series();
+            series.setName(key);
+            seriesMap.put(key, series);
+        }
+        return seriesMap.get(key);
+    }
+
+    private String getBuildOn(JenkinsBuild build) {
+        return build.getBuiltOn() == null ? "unknown" : build.getBuiltOn();
+    }
+
+
 }
