@@ -4,12 +4,11 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.jenkinsutil.payload.in.JenkinsBuildDistribution;
 import de.jenkinsutil.payload.mapping.BuildDistributionMapper;
-import de.jenkinsutil.payload.out.BuildDay;
+import de.jenkinsutil.payload.out.SeriesHolder;
 import org.springframework.stereotype.Service;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -21,20 +20,14 @@ public class BuildDistributionService {
         this.buildDistributionMapper = buildDistributionMapper;
     }
 
-    public Optional<List<BuildDay>> getBuildDays() {
+    public Optional<SeriesHolder> getBuildDays() throws IOException {
+        JenkinsBuildDistribution jenkinsBuildDistribution = readBuildDistribution();
+        return Optional.of(buildDistributionMapper.convert(jenkinsBuildDistribution));
+    }
 
+    private JenkinsBuildDistribution readBuildDistribution() throws IOException {
         File jsonFile = new File("/home/manuel/Dokumente/jenkins_build_overview.json");
-
         ObjectMapper mapper = new ObjectMapper().configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-        try {
-            JenkinsBuildDistribution jenkinsBuildDistribution = mapper.readValue(jsonFile, JenkinsBuildDistribution.class);
-            return Optional.of(buildDistributionMapper.convert(jenkinsBuildDistribution));
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        // TODO: implement
-        return Optional.empty();
+        return mapper.readValue(jsonFile, JenkinsBuildDistribution.class);
     }
 }
